+++ Source/Streamstone/Event.cs
@@ -91,7 +91,7 @@ namespace Streamstone
             Properties = properties;
         }
 
-        internal RecordedEvent Record(Partition partition, long version) =>
+        internal RecordedEvent Record(Partition partition, int version) =>
             new RecordedEvent(Id, Properties, Includes, partition, version);
     }
 
@@ -114,12 +114,12 @@ namespace Streamstone
         /// <summary>
         /// A sequence number assigned by a stream to this event. 
         /// </summary>
-        public long Version { get; }
+        public int Version { get; }
 
         internal readonly EntityOperation[] EventOperations;
         internal readonly EntityOperation[] IncludedOperations;
 
-        internal RecordedEvent(EventId id, EventProperties properties, IEnumerable<Include> includes, Partition partition, long version)
+        internal RecordedEvent(EventId id, EventProperties properties, IEnumerable<Include> includes, Partition partition, int version)
         {
             Id = id;
             Version = version;
